extrev: add factory function, move parent class to base module
