+++ python/nistoar/midas/dap/extrev/__init__.py
@@ -2,61 +2,47 @@
 a subpackage that defines an interface and implementations for interacting with external 
 review systems
 
-The :py:class:`ExternalReviewClient` defines the interface.
+The :py:class:`~nistoar.midas.dap.extrev.base.ExternalReviewClient` base class defines 
+the interface.
 """
-from abc import ABC, abstractmethod
+from .base import *
+from .nps1 import NPSExternalReviewClient
+# from .sim import SimulatedExternalReviewClient
+from nistoar.base.config import ConfigurationException
 
-from nistoar.midas import MIDASException
+__all__ = [ "ExternalReviewClient", "ExternalReviewException", "create_external_review_client" ]
 
-__all__ = [ "ExternalReviewClient", "ExternalReviewException" ]
+_client_classes = {
+    NPSExternalReviewClient.system_name:        NPSExternalReviewClient
+#   SimulatedExternalReviewClient.system_name:  SimulatedExternalReviewClient
+}
 
-class ExternalReviewException(MIDASException):
+def create_external_review_client(config: Mapping) -> ExternalReviewClient:
     """
-    a base exception for issues when interacting with an external review system
+    a factory function that creates a client for submitting DAP records for external review.  
+
+    This is called by :py:meth:`create_service_for` to inject an 
+    :py:class:`~nistoar.midas.dap.extrev.ExternalReviewClient` into its 
+    :py:class:`DAPService`.  If ``config`` is not give or is empty, None is returned;
+    this will cause external review to be disabled and not required for record publication.
+
+    If non-empty, the given configuration must include the ``name`` parameter, which identifies
+    which class will be instantiated.  All other configuration paramters are class-dependent.
+
+    :raises ConfigurationException: if the given, non-empty configuration lacks a ``name`` or its
+                                    value is unrecognized.
+    parameter.  
     """
-    pass
+    if not config:
+        return None
 
-class ExternalReviewClient(ABC):
-    """
-    an interface for interacting with a remote DAP review system.  
-    """
-    system_name = "unspecified"
+    if not config.get('name'):
+        raise ConfigurationException("external_review: missing required name parameter")
 
-    def __init__(self, config=None):
-        """
-        initialize the client
-        """
-        self.cfg = config or {}
+    cls = _client_classes.get(config['name'])
+    if not cls:
+        raise ConfigurationException("external_review: name not supported: "+config['name'])
 
-    @abstractmethod
-    def submit(self, id: str, submitter: str, version: str=None, **options):
-        """
-        submit a specified DAP to this system for review.  The options supported depend on 
-        the implementation (and implementations should ignore any option parameters that 
-        it does not recognize; however, support for the following parameters are defined as 
-        follows:
-        ``instructions``
-             (*[str]*) a list of statements that should be passed as special instructions to 
-             the reviewers.
-        ``changes``
-             (*[str]*) provided when a DAP revision is being requested for review, this is a 
-             list of statements indicating what has changed since the previous publication.
-        ``reviewers``
-             (*[dict]*) a list of designations of reviewers requested to be included 
-             among the full set of assigned reviewers.  
-        ``security_review``
-             (*bool*) if True, this DAP includes content (e.g. software) that requires 
-             review IT security review.  
-
-        :param str id:         the identifier for the DAP to submit to the review system
-        :param str submitter:  the identifier for the user submitting the DAP (this is 
-                               usually the owner of the record).
-        :param str version:    the version of the DAP being submitted.  This should be 
-                               provided when revising a previously published DAP; otherwise,
-                               the implementation may assume that the initial version is being 
-                               submitted.  
-        :param Mapping options:  extra implementation-specific keyword parameters
-        """
-        raise NotImplemented()
+    return cls(config)
